+++ tests/testthat/test-tune_covlmc.R
@@ -100,3 +100,33 @@ test_that("tune_vlmc verbosity is adequate", {
   df_y <- data.frame(y = y, z = runif(length(y)))
   expect_snapshot_output(tune_covlmc(x, df_y, criterion = "BIC", verbose = 1))
 })
+
+test_that("tune_covlmc trimming", {
+  withr::local_seed(0)
+  x <- sample(c("A", "B", "C"), 1000, replace = TRUE)
+  y <- ifelse(runif(length(x)) > 0.5, c(x[-1], sample(c("A", "B", "C"), 1)), c(x[-c(1, 2)], sample(c("A", "B", "C"), 2, replace = TRUE)))
+  y <- as.factor(ifelse(runif(length(x)) > 0.2, y, sample(c("A", "B", "C"), 500, replace = TRUE)))
+  df_y <- data.frame(y = y, z = runif(length(y)))
+  bt_covlmc <- tune_covlmc(x, df_y,
+    criterion = "BIC", save = "all", trimming = "partial",
+    best_trimming = "partial"
+  )
+  expect_equal(length(bt_covlmc$saved_models$all) + 1L, nrow(bt_covlmc$results))
+  best_BIC_idx <- which.min(bt_covlmc$results$BIC)
+  expect_identical(bt_covlmc$best_model, bt_covlmc$saved_models$all[[best_BIC_idx - 1]])
+  for (model in bt_covlmc$saved_models$all) {
+    expect_no_error(contexts(model, hsize = TRUE, model = "full"))
+  }
+  expect_no_error(contexts(bt_covlmc$saved_models$initial, hsize = TRUE, model = "full"))
+  bt_covlmc <- tune_covlmc(x, df_y,
+    criterion = "BIC", save = "all", trimming = "full",
+    best_trimming = "full"
+  )
+  expect_equal(length(bt_covlmc$saved_models$all) + 1L, nrow(bt_covlmc$results))
+  best_BIC_idx <- which.min(bt_covlmc$results$BIC)
+  expect_identical(bt_covlmc$best_model, bt_covlmc$saved_models$all[[best_BIC_idx - 1]])
+  for (model in bt_covlmc$saved_models$all) {
+    expect_error(contexts(model, hsize = TRUE, model = "full"))
+  }
+  expect_error(contexts(bt_covlmc$saved_models$initial, hsize = TRUE, model = "full"))
+})